Implemented and tested the fast algorithm

+++ collinear/src/main/java/Point.java
@@ -65,9 +65,9 @@ public class Point implements Comparable<Point> {
         if (this.x == that.x)
             return Double.POSITIVE_INFINITY;
         if (this.y == that.y)
-            return 0;
+            return 0.0;
 
-        return (that.y - this.y) / (that.x - this.x);
+        return (double) (that.y - this.y) / (that.x - this.x);
     }
 
     /**
